# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "蘑菇博客搜索服务",
        "title": "蘑菇博客搜索服务",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
        "/search/addElasticSearchIndexByUid": {
            "post": {
                "description": "通过博客Uid添加索引",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "索引相关"
                ],
                "summary": "通过博客Uid添加索引",
                "parameters": [
                    {
                        "type": "string",
                        "description": "博客uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加提示",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/search/deleteElasticSearchByUid": {
            "post": {
                "description": "删除博客索引",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "索引相关"
                ],
                "summary": "删除博客索引",
                "parameters": [
                    {
                        "type": "string",
                        "description": "删除uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除提示",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/search/deleteElasticSearchByUids": {
            "post": {
                "description": "批量删除博客索引",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "索引相关"
                ],
                "summary": "批量删除博客索引",
                "parameters": [
                    {
                        "type": "string",
                        "description": "删除uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除提示",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/search/elasticSearchBlog": {
            "get": {
                "description": "博客搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "索引相关"
                ],
                "summary": "博客搜索",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "currentPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "querystring",
                        "name": "searchType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.DocIndex"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/search/initElasticSearchIndex": {
            "post": {
                "description": "初始化索引",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "索引相关"
                ],
                "summary": "初始化索引",
                "responses": {
                    "200": {
                        "description": "初始化索引提示",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.DocIndex": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string"
                },
                "blogSortName": {
                    "type": "string"
                },
                "blogSortUid": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isPublish": {
                    "type": "string"
                },
                "oid": {
                    "type": "integer"
                },
                "photoUrl": {
                    "type": "string"
                },
                "sortName": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tagNameList": {
                    "description": "标签名称",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tagUidList": {
                    "description": "标签id",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "资源类型: 文章、动态、问答、面经",
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "response.JsonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        }
    }
}